+++ src/roxbot/gps/converters.py
@@ -13,6 +13,21 @@ from typing import Tuple
 from pymap3d import enu2geodetic, geodetic2enu  # type: ignore
 
 
+GPS_REF: tuple[float, float] | None = None
+
+# get gps reference from environment variable
+var = os.environ.get("GPS_REF")
+
+if var is not None:
+    GPS_REF = tuple(float(x) for x in var.split(","))  # type: ignore
+
+
+def set_gps_ref(lat: float, lon: float) -> None:
+    """set gps reference point"""
+    global GPS_REF  # pylint: disable=global-statement
+    GPS_REF = (lat, lon)
+
+
 def heading_to_theta(angle_deg: float) -> float:
     """convert gps heading to theta in radians"""
     h = -1j * cmath.rect(1, radians(angle_deg))
@@ -25,11 +40,29 @@ def theta_to_heading(angle_rad: float) -> float:
     return degrees(-cmath.phase(h))  # type: ignore
 
 
+def latlon_to_enu(latlon: Tuple[float, float]) -> Tuple[float, float]:
+    if GPS_REF is None:
+        raise ValueError("GPS_REF not set")
+    x, y, _ = geodetic2enu(latlon[0], latlon[1], 0, GPS_REF[0], GPS_REF[1], 0)
+    return float(x), float(y)
+
+
+def enu_to_latlon(xy: Tuple[float, float]) -> Tuple[float, float]:
+    if GPS_REF is None:
+        raise ValueError("GPS_REF not set")
+    lat, lon, _ = enu2geodetic(xy[0], xy[1], 0, GPS_REF[0], GPS_REF[1], 0)
+    return float(lat), float(lon)
+
+
 class GpsConverter:
     """coordinate converter (lat,lon) to (xEast,yNorth). wrapper around pymap3d"""
 
     def __init__(self, gps_ref: tuple[float, float] | None = None) -> None:
         """create converter, try to get reference from GPS_REF env variable"""
+        warnings.warn(
+            "GpsConverter is deprecated, no longer needed. Use functions from converters directly."
+        )
+
         if gps_ref is not None:
             self.gps_ref = gps_ref
         else: